+++ src/babashka/impl/classpath.clj
@@ -84,8 +84,8 @@
       cp)))
 
 (defn resource
-  (^URL [path] (resource @the-url-loader path))
-  (^URL [loader path]
+  (^URL [path] (resource path @the-url-loader))
+  (^URL [path loader]
    (if (str/starts-with? path "/") nil ;; non-relative paths always return nil
        (getResource loader [path] true))))