+++ CHANGELOG.md
@@ -18,6 +18,8 @@ A preview of the next release can be installed from
 - [#1822](https://github.com/babashka/babashka/issues/1822): type should prioritize :type metadata
 - `ns-name` should work on symbols
 - `:clojure.core/eval-file` should affect `*file*` during eval
+- [#1179](https://github.com/babashka/babashka/issues/1179): run `:init` in tasks only once
+- [#1823](https://github.com/babashka/babashka/issues/1823): run `:init` in tasks before task specific requires
 
 ## 1.12.200 (2025-04-26)