+++ CHANGELOG.md
@@ -17,6 +17,7 @@ A preview of the next release can be installed from
 - Bump `fs` to `0.5.25`
 - Bump `jsoup` to `1.20.1`
 - Bump `edamame` to `1.4.30`
+- Bump `taoensso.timbre` to `6.7.0`
 - [#1815](https://github.com/babashka/babashka/issues/1815): Make install-script wget-compatible ([@eval](https://github.com/eval))
 - [#1822](https://github.com/babashka/babashka/issues/1822): type should prioritize :type metadata
 - `ns-name` should work on symbols