+++ CHANGELOG.md
@@ -19,6 +19,7 @@ A preview of the next release can be installed from
 - Bump `jsoup` to `1.20.1`
 - Bump `edamame` to `1.4.30`
 - Bump `taoensso.timbre` to `6.7.0`
+- Bump `pods`: more graceful error handling when pod quits unexpectedly
 - [#1815](https://github.com/babashka/babashka/issues/1815): Make install-script wget-compatible ([@eval](https://github.com/eval))
 - [#1822](https://github.com/babashka/babashka/issues/1822): `type` should prioritize `:type` metadata
 - `ns-name` should work on symbols